move mongodb core folder
