+++ ubotindo/modules/no_sql/gban_db.py
@@ -15,7 +15,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """Global bans utils."""
 
-from ubotindo.utils import get_collection
+from ubotindo.modules.no_sql import get_collection
 
 GBAN_USER = get_collection("GBANS")
 GBAN_SETTINGS = get_collection("GBAN_SETTINGS")
