--- ubotindo/utils/__init__.py
@@ -1,1 +1,0 @@
-from .database import get_collection
